==> figures/2_sinkhorn_visualization.Rmd <==
---
title: "Figure 2: Sinkhorn Visualization"
output: html_document
---

```{r}
source('../R/setup.R')
source("../R/plotting_methods.R")
library(plotly)
library(RColorBrewer)
```

```{r}
set.seed(4)

N <- 500
M <- 400
K <- 3
```

# Generate data
```{r}
H <- matrix(runif(K * N, min = 0, max = 10), nrow = K )  
W <- t(matrix(runif(K * M, min = 0, max = 10), nrow = K ) )
rownames(W) <- c(1:dim(W)[[1]])
colnames(H) <- c(1:dim(H)[[2]])
V <- W %*% H
V_no_noise <- V
```



# Proper sinkhorn

## Colored
```{r fig.height = 5, fig.width = 10}
dso <-  DualSimplexSolver$new()
dso$set_data(V, max_sinkhorn_iterations = 20)
dso$project(K)
dso$plot_projected("plane_distance", "plane_distance", with_solution=TRUE, use_dims = list(2:3))
```


## 3D 
````{r fig.height = 20, fig.width = 20}
points_col <- brewer.pal(9, "Blues")
plot_3d(round(dso$st$proj$X, 5), "R", size = 3, colors = points_col, shift=0)
points_col <- brewer.pal(9, "Oranges")
plot_3d(round(dso$st$proj$Omega, 5), "S", size = 3,  colors = points_col, shift=0)
```


# Check the effect of normalization

## Original space no simplexes (centered data)
```{r}
data <- V
centered_data <- data - rowMeans(data) # center the data

# now want to calculate SVD
svd_ <- svd(centered_data)
S <- t(svd_$u[, 1:2])
R <- t(svd_$v[, 1:2])

V_r <-  t(R %*% t(centered_data))
V_s <-  t(S %*% centered_data)

proj <- list(
  X = V_r,
  Omega = V_s,
  meta = list(K = K, N = N, M = M)
)
plot <- plot_projection_points(proj, use_dims = c(1, 2))
plot
```

## 3D svd with next component
```{r}
S <- t(svd_$u[, 1:3])
R <- t(svd_$v[, 1:3])

V_r <-  t(R %*% t(centered_data))
V_s <-  t(S %*% centered_data)
points_col <- brewer.pal(9, "Blues")
plot_3d(V_r, "V", size = 6, colors = points_col)
points_col <- brewer.pal(9, "Oranges")
plot_3d(V_s, "V", size = 6,  colors = points_col)
```

## 3D SVD without centering
```{r}
data <-   V
#centered_data <-  data - rowMeans(data) # center the data

# now want to calculate SVD
svd_ <- svd(data)
S <- t(svd_$u[, 1:K])
R <- t(svd_$v[, 1:K])

V_r <-  t(R %*% t(data))
V_s <-  t(S %*% data)


plot_3d(V_r, "R", size = 6)
plot_3d(V_s, "S", size = 6)
```


## Row normalization

### Centered data
```{r}
data <- V
norm_data <- data / rowSums(data)
centered_data <- t(t(norm_data) - colMeans(norm_data)) # center the data

# now want to calculate SVD
svd_ <- svd(centered_data)
S <- t(svd_$u[, 1:2])
R <- t(svd_$v[, 1:2])

V_r <-  t(R %*% t(centered_data))
V_s <-  t(S %*% centered_data)

proj <-  list(
  X = V_r,
  Omega = V_s,
  meta = list(K = K - 1, N = N, M = M)
)
plot <- plot_projection_points(proj, use_dims = c(1, 2))
plot
```

### 3D svd with next component (centered data)
```{r}
S <- t(svd_$u[, 1:K])
R <- t(svd_$v[, 1:K])

V_r <-  t(R %*% t(centered_data))
V_s <-  t(S %*% centered_data)
points_col <- brewer.pal(9, "Blues")
plot_3d(V_r, "V", size = 6, colors = points_col)
points_col <- brewer.pal(9, "Oranges")
plot_3d(V_s, "V", size = 6,  colors = points_col)
```


### 3D picture without centering
```{r}
svd_ <- svd(norm_data)
S <- t(svd_$u[, 1:K])
R <- t(svd_$v[, 1:K])

V_r <-  t(R %*% t(norm_data))
V_s <-  t(S %*% norm_data)
points_col <- brewer.pal(9, "Blues")
plot_3d(V_r, "V", size = 6, colors = points_col)
points_col <- brewer.pal(9, "Oranges")
plot_3d(V_s, "V", size = 6,  colors = points_col)
```


## Column normalization
```{r}
data <- V
norm_data <- t(t(data) / colSums(data))
centered_data <- (norm_data) - rowMeans(norm_data) # center the data
centered_data <-
  t(t(norm_data) - colMeans(norm_data)) # center the data

# now want to calculate SVD
svd_ <- svd(centered_data)
S <- t(svd_$u[, 1:2])
R <- t(svd_$v[, 1:2])

V_r <-  t(R %*% t(centered_data))
V_s <-  t(S %*% centered_data)

proj <-  list(X = V_r,
              Omega = V_s,
              meta = list(K = K - 1, N = N, M = M))
plot <- plot_projection_points(proj, use_dims = c(1, 2))
plot
```

### 3D svd with next component (centered data)
```{r}
S <- t(svd_$u[, 1:3])
R <- t(svd_$v[, 1:3])

V_r <-  t(R %*% t(centered_data))
V_s <-  t(S %*% centered_data)
points_col <- brewer.pal(9, "Blues")
plot_3d(V_r, "V", size = 6, colors = points_col)
points_col <- brewer.pal(9, "Oranges")
plot_3d(V_s, "V", size = 6,  colors = points_col)
```
