==> figures/4_1_nmf_setup_ensure_working_with_simulated_data.Rmd <==
---
title: "Figure 5: NMF, Part 1."
output: html_notebook
---


```{r}
dir_to_save_fig <- "../out/4_NMF/"
dir.create(file.path(".", dir_to_save_fig), showWarnings = F, recursive = T)
source('../R/setup.R')
source('../R/nmf_methods.R')
```


# Syntetic experiment 
Here we use Syntetic data from NMF package just to visualize and 
check that everything works fine

## Make the NMF Data matrix
```{r}
components_count <- c(3, 4, 5)
mixtures_count <- c(25, 50, 100, 500)
features_count <- c(200, 500, 1000, 2000)
```

```{r}
current_K <- components_count[[2]]
current_M <- features_count[[2]]
current_N <- mixtures_count[[1]]
number_of_runs <- 5
```

```{r}
global_generated_data <- make_syntetic_data(M = current_M, N = current_N, K = current_K)
```

```{r}
plot_w_h_v(global_generated_data$true_W, global_generated_data$true_H, global_generated_data$V)
```


## Solve this NMF using the NMF package (3 different algorithms here)
```{r}
res_methods <-
  NMF::nmf(
    global_generated_data$V,
    current_K,
    list('lee', 'brunet', 'nsNMF'),
    nrun = number_of_runs,
    .options = 'kt'
  )
```


```{r}
res_halsAcc <-
  NMF::nmf(
    global_generated_data$V,
    current_K,
    hNMF::HALSacc,
    nrun = number_of_runs,
    .options = 'tk',
    name = "HALSacc"
  )
res_methods[["HALSacc"]] <- res_halsAcc
res_PGNMF <-
  NMF::nmf(
    global_generated_data$V,
    current_K,
    hNMF::PGNMF,
    nrun = number_of_runs,
    .options = 'tk',
    name = "PGNMF"
  )
res_methods[["PGNMF"]] <- res_PGNMF
res <-
  NMF::nmf(
    global_generated_data$V,
    current_K,
    als_nmf_algorithm,
    nrun = number_of_runs,
    .options = 'tk',
    name = "ALS"
  )
res_methods[["ALS"]] <- res
```


# Solve the same NMF problem using Dual Simplex Method
```{r}
res <-
  NMF::nmf(
    global_generated_data$V,
    current_K,
    dualsimplex_nmf_algorithm,
    nrun = number_of_runs,
    .options = 'tk',
    name = "DualSimplex"
  )
res_methods[["DualSimplex"]] <- res
```


## Check results
### Triangles found by DualSimplex method
```{r}
check_triangles(res_methods)
```


### Matrices for each algorithm
Plot W, H , V for each algorithm used
```{r}
lapply(res_methods, function (method_result_list){
  method_result <- method_result_list[[1]] 
  current_W <- method_result@fit@W
  current_H <- method_result@fit@H
  plot_w_h_v(current_W, current_H, current_W %*% current_H, method_result@method)
})
```

### RMSE errors

```{r, fig.height = 5, fig.width=20}
data_used <- rep(list(global_generated_data), number_of_runs)
distances_to_check <- plot_metrics_box(res_methods = res_methods, data_used=data_used, folder_to_save_results = dir_to_save_fig)
distances_to_check$plot
```
