==> figures/5b_gse11058.Rmd <==
---
title: "Figure 6A: Deconvolution of GSE11058"
output: html_document
---

```{r}
# path to save everything
dir_to_save_fig <- "../out/5_GSE11058/"
dir.create(file.path(".", dir_to_save_fig), showWarnings = F, recursive = T)
source("../R/figure_utils.R") # for same colors
source('../R/setup.R') # import main package
```


## Dependencies
```{r}
if (!require("hgu133plus2.db", quietly = TRUE))
  BiocManager::install("hgu133plus2.db")
if (!require("ComplexHeatmap", quietly = TRUE))
  BiocManager::install("ComplexHeatmap")
if (!require("nnls", quietly = TRUE))
  install.packages("nnls")
if (!require("irlba", quietly = TRUE))
  install.packages("irlba")
if (!require("reshape", quietly = TRUE))
  install.packages("reshape")
if (!require("linseed", quietly = TRUE))
  devtools::install_github("ctlab/linseed")
if (!require("biomaRt", quietly = TRUE))
  BiocManager::install("biomaRt")
if (!require("ggrastr", quietly = TRUE))
  install.packages("ggrastr")
if (!require("svglite", quietly = TRUE))
  install.packages("svglite")


library('GEOquery')
library(biomaRt)
library(hgu133plus2.db)
library(ComplexHeatmap)
library(progress)
library(reshape)
library(plotly)
library(ggrastr)
library(ggpubr)
library(linseed)
library(svglite)
library(DualSimplex)
```


# Prepare Data

## Get it from GEO
```{r}
n_ct <- 4

library('GEOquery')
dataset <- "GSE11058"
gse <- getGEO(dataset, AnnotGPL = T)
gse <- gse[[1]]
data_raw <- Biobase::exprs(gse)
#linearize the data!
data_raw <- linearize_dataset(data_raw)
```


## Map probes to genes

probe id will be converted and collapsed to gene names

```{r}
# probes which are present in data
probe_mapping <-
  biomaRt::select(hgu133plus2.db::hgu133plus2.db,
                  rownames(data_raw),
                  c("SYMBOL"),
                  keytype = "PROBEID")
probe_mapping <-
  probe_mapping[probe_mapping$PROBEID %in% rownames(data_raw),]
# not empty result gene id (now we  many-to-many in in genes to probes)
probe_mapping <- probe_mapping[!is.na(probe_mapping$SYMBOL),]
## want to collapse gene for each gene and take mean value of data
data_raw <- data_raw[probe_mapping$PROBEID, ]
rownames(data_raw) <- probe_mapping$SYMBOL
# should be unique gene name as row name after this
data_raw <-
  tapply(data_raw, list(row.names(data_raw)[row(data_raw)], colnames(data_raw)[col(data_raw)]), FUN = median)
```


## Extract true proportions (H) and basis (W) from the data
Info about true samples is included for this dataset. We use this info for validation.

### True H

```{r}
pdata <- pData(gse)
component_names <- c("Jurkat", "IM-9", "Raji", "THP-1")

mix_a <- c(2.5, 1.25, 2.5, 3.75)
mix_a <- mix_a / sum(mix_a)
mix_b <- c(0.5, 3.17, 4.75, 1.58)
mix_b <- mix_b / sum(mix_b)
mix_c <- c(0.1, 4.95, 1.65, 3.3)
mix_c <- mix_c/ sum(mix_c)
mix_d <- c(0.02, 3.33, 3.33, 3.33)
mix_d <- mix_d / sum(mix_d)

proportions <- matrix(0 , nrow = dim(data_raw)[[2]], ncol = n_ct )

proportions[1:3, 1] <- 1
proportions[4:6, 2] <- 1
proportions[7:9, 3] <- 1
proportions[10:12, 4] <- 1

proportions[13, ] <- mix_a
proportions[14, ] <- mix_a
proportions[15, ] <- mix_a

proportions[16, ] <- mix_b
proportions[17, ] <- mix_b
proportions[18, ] <- mix_b

proportions[19, ] <- mix_c
proportions[20, ] <- mix_c
proportions[21, ] <- mix_c

proportions[22, ] <- mix_d
proportions[23, ] <- mix_d
proportions[24, ] <- mix_d

rownames(proportions) <- rownames(pdata)
colnames(proportions) <- component_names
true_proportions <- t(proportions)
```

### True W
```{r}
components <- lapply(c(1:n_ct), function(comp_num){
  component_subset_columns <- colnames(true_proportions[,true_proportions[comp_num, ] == 1])
  component_subset <- data_raw[, component_subset_columns]
  component_vector <- as.matrix(rowMedians(component_subset))
  rownames(component_vector) <- rownames(component_subset)
  return(component_vector)

})
true_basis <- do.call(cbind, components)
colnames(true_basis) <- component_names
true_marker_list <- get_signature_markers(true_basis, n_marker_genes = 100)

```

## Remove pure samples
(Optional, to make the task harder)
```{r}
#sample_is_mixed <- apply(true_proportions, 2, mad) != 0
#true_proportions <- true_proportions[, sample_is_mixed]
#data_raw <- data_raw[, sample_is_mixed]
```

## Print data dimensions
```{r}
print(paste("Dim for the whole data:", toString(dim(data_raw))))
print(paste("Dim for hidden proportions (H):", toString(dim(true_proportions))))
print(paste("Dim for hidden basis (W):", toString(dim(true_basis))))
```

# Dual simplex method

## Create lo2 object and prefilter

```{r}
dso <- DualSimplexSolver$new()
dso$set_data(data_raw)
# remove "RPLS", "LOC", "ORF", "SNOR", keep coding only
dso$basic_filter()
dso$project(n_ct)
print(paste("Dim for the prefiltered data:", toString(dim(dso$get_data()))))
```

```{r fig.height = 5, fig.width = 11}
dso$plot_svd(dims = 1:15)
dso$plot_projected("zero_distance", "zero_distance", use_dims = list(3:4))
```

## Additional filters

### MAD filtering

Check distribution of median absolute deviation to see which genes will be deleted
```{r}
dummy_threshold <- 2.5
data <- dso$get_data()
anno <- get_anno(data)
anno$PASS_FILTER <- FALSE
anno[anno$log_mad > dummy_threshold,]$PASS_FILTER <- TRUE
data <- set_anno(anno, data)
plot_feature(data, feature = "log_mad", col_by ='PASS_FILTER')
```

Filter the data with selected threshold
```{r fig.height = 4, fig.width=14}
dso$basic_filter(log_mad_gt = dummy_threshold, )
dso$project(n_ct)

svd_plot_2 <- dso$plot_svd(1:10) + theme_minimal(base_size = 8)
points_2 <- dso$plot_projected("black", "black",use_dims = 2:3, show_plots=F) +  theme_minimal(base_size = 8)
plotlist = list(points_2, svd_plot_2)
cowplot::plot_grid(plotlist=plotlist, rel_widths = c(0.66, 0.33))
```

Repeat previous cell with increasing thresholds to observe how SVD singular values will change. (in a perfect case should see only n_ct non-zero singular values)
```{r}
dso$plot_svd_history()
```


### Denoising the data

Plot distance of genes from the projection hyperplane
```{r fig.height = 4, fig.width=14}
distance_dist_plot_noizy <-
  plot_feature(dso$get_data(), feature = "plane_distance") + theme_minimal(base_size = 15)
color = fData(dso$get_data())$plane_distance
X_points_noizy <-
  plot_projection_points(
    dso$st$proj,
    use_dims = 2:3,
    color = color,
    color_name = "plane_distance",
    spaces = c("X"),
    pt_size = 2
  ) + theme_minimal(base_size = 15)

distances_distribution <-
  plot_feature_pair(dso$get_data(), "plane_distance", "zero_distance", T, size = 0.1)


plotlist = list(X_points_noizy,
                distance_dist_plot_noizy,
                distances_distribution)

cowplot::plot_grid(
  plotlist = plotlist,
  nrow = 1,
  rel_widths = c(0.4, 0.3, 0.3)
)
```

Remove genes which are to far from projection hyperplane and zero point
```{r}
dso$distance_filter(plane_d_lt = 0.2, zero_d_lt = 0.3, genes = T)
dso$project(n_ct)
```

Color points by distance
```{r fig.height = 4, fig.width=14}
current_svd_plot <- dso$plot_svd(1:10) + theme_minimal(base_size = 8)
current_points_plot <-
  dso$plot_projected("plane_distance",
                     "plane_distance",
                     use_dims = 2:3,
                     show_plots = F) + theme_minimal(base_size = 8)
plotlist = list(current_points_plot, current_svd_plot)
cowplot::plot_grid(plotlist = plotlist, rel_widths = c(0.66, 0.33))
```

Repeat previous cell with decreasing thresholds to observe how SVD singular values and coloring will change. (in a perfect case should see only n_ct non-zero singular values)
```{r}
dso$plot_svd_history()
```


## Make umap in advance for visualization purposes
since we have 3-dimensional simplex it is hard to visualize it in 2D. In practice UMAP looks good if you want to estimate the number of simplex corners and track the solution
```{r}
set.seed(23)
dso$run_umap(neighbors_X = 25, neighbors_Omega = 5)
dso$plot_projected("zero_distance", "zero_distance", use_dims = NULL)
```


## Finaly train the model
### Initialze random points
```{r}
#set.seed(72)
set.seed(33)
dso$init_solution("random")
# original space
dso$plot_projected("zero_distance", "zero_distance", use_dims = c(3:4))
# umap visualization
dso$plot_projected("zero_distance", "zero_distance", use_dims = NULL)
```

### Train multiple times with changing hyperparameters
We make the learning rate smaller for later iterations.
We also increase the contribution of term related to positivity constraint for later iterations.
```{r}
# This is default optimization method which worth to try in all cases.
dso$default_optimization()

# Below is a rough content of this method for those who want to tune params (see below and supplementary notes on guidance)
# LR_DECAY_STEPS = 1
# PARAMETERS_INCREASE_STEPS = 4
# lr_decay <- 0.5
# params_increase <- 2
# original_lambda_term <- 0.01  #coef_hinge_H
# original_beta_term <- 0.1 #coef_hinge_W
# lr_x <- 0.001
# lr_omega <- 0.001
# 
# RUNS_EACH_STEP <- 5000
# 
# pb <- progress_bar$new(total = LR_DECAY_STEPS * PARAMETERS_INCREASE_STEPS)
# 
# for (lr_step in 1:LR_DECAY_STEPS) {
#   lambda_term <- original_lambda_term
#   beta_term <- original_beta_term
#   for (x in 1:PARAMETERS_INCREASE_STEPS) {
#     pb$tick()
#     dso$optim_solution(
#       RUNS_EACH_STEP,
#       optim_config(
#         coef_hinge_H = lambda_term,
#         coef_hinge_W = beta_term,
#         coef_der_X = lr_x,
#         coef_der_Omega = lr_omega
#       )
#     )
#     lambda_term <- lambda_term * params_increase
#     beta_term <- beta_term * params_increase
# 
#   }
#   lr_x <- lr_x * lr_decay
#   lr_omega <- lr_omega * lr_decay
# }
```


## Error history
As we describe in the paper -- we have multiple error terms.
Behavior of which are controlled by parameters.
Values coef_der_X and coef_der_Omega should be understood as learning rate and and lower values will lead to smaller steps of optimization. (they ae mu and nu from the paper)

Values coef_hinge_H and coef_hinge_W (Lambda, Beta) are controlling the magnitude for negativity terms for matrix H and W respectively in the error.
Since we have multiple terms we need to ensure that all terms are converging. 
As we can see from this particular example: deconvolution term goes down first. Algorithm does not care about negativity which leads to increase in number of negative elements in the solution. But once deconvolution error is small enough algorithm will start opimizing negativity as well. 

The idea in general is to keep error terms close to each other to allow optimization of all terms. You can see suplementary notes script for more details

```{r}
dso$plot_error_history()
dso$plot_negative_basis_change()
dso$plot_negative_proportions_change()
```


## Prepare plots as they are in the paper
```{r}
# This option makes all dots as a single layer reducing result image size
options("dualsimplex-rasterize" = T)
```

### Save errors plot with rasterization
```{r}
plotErrorsWithRastr = function(metadata,
                               variables = c(
                                 "deconv_error",
                                 "lamdba_error",
                                 "beta_error",
                                 "D_h_error",
                                 "D_w_error",
                                 "total_error"
                               )) {
  solution_proj <- metadata$st$solution_proj
  error_statistics <- solution_proj$optim_history$errors_statistics

  toPlot <- data.frame(error_statistics[, variables])

  toPlot$iteration <- 0:(nrow(error_statistics) - 1)
  toPlot <-
    melt(toPlot, id.vars = "iteration", measure.vars = variables)
  plt <-
    ggplot(toPlot, aes(
      x = iteration,
      y = log10(value),
      color = variable
    )) +
    # rasterise(geom_point(size=0.2),dpi=600) +
    rasterise(geom_line(size = 0.8), dpi = 600) + theme_minimal() + labs(color =
                                                                           "Errors")
  return(plt)
}

errors_plot <-
  plotErrorsWithRastr(dso,
                      variables = c("deconv_error",
                                    "lamdba_error",
                                    "beta_error",
                                    "total_error")) +
  theme(text = element_text(size = 16))

filename <- paste0(dir_to_save_fig, "simulation_trajectory", ".svg")
ggsave(
  file = filename,
  plot = errors_plot,
  width = 5,
  height = 3,
  device = svglite
)
errors_plot
```


### History of solution
```{r fig.width=5, fig.height=5}
colnames(dso$st$proj$X) <- c("R1", "R2", "R3", "R4")
colnames(dso$st$proj$Omega) <- c("S1", "S2", "S3", "S4")
colnames(dso$st$proj$umap$X)  <- c("UMAP1", "UMAP2")
colnames(dso$st$proj$umap$Omega)  <- c("UMAP1", "UMAP2")

proj_solution_history <- dso$plot_projected(use_dims = (2:3), wrap=F, with_legend = F)
proj_solution_history[[1]] # genes in a space of samples
proj_solution_history[[2]] # samples in a space of genes
```


Use UMAP and decorate it how it was done in the paper
```{r}
umap_solution_history <-
  dso$plot_projected(
    "zero_distance",
    "zero_distance",
    wrap = F,
    with_legend = F,
    use_dims = NULL
    #, from_iter = 5000
  )
pltX <- umap_solution_history[[1]] +
  scale_fill_manual(values = colors_v[3:6]) +
  theme_minimal(base_size = 12)

pltX_no_leg <- pltX +
  theme(legend.position = "none")
filename <-
  paste0(dir_to_save_fig, "proj_solution_history_X", ".svg")
ggsave(
  file = filename,
  plot = pltX_no_leg,
  width = 3,
  height = 3,
  device = svglite
)

filename <-
  paste0(dir_to_save_fig, "proj_solution_history_X_with_legend", ".svg")
ggsave(
  file = filename,
  plot = pltX,
  width = 3,
  height = 3,
  device = svglite
)

pltOmega <- umap_solution_history[[2]] +
  scale_fill_manual(values = colors_v[3:6]) +
  theme_minimal(base_size = 12) +
  theme(legend.position = "none")
filename <-
  paste0(dir_to_save_fig, "proj_solution_history_Omega", ".svg")
ggsave(
  file = filename,
  plot = pltOmega,
  width = 3,
  height = 3,
  device = svglite
)
ggarrange(pltX, pltOmega)
```


### Solution check plots
Get solution (W and H)
```{r}
solution <- dso$finalize_solution()
```


#### Compare with true proportions/basis
Check ratio of true proportion or component expression to predicted
```{r fig.width=20, fig.height=5}
ptb <- coerce_pred_true_basis(solution$W, true_basis[rownames(solution$W), ])
ptp <- coerce_pred_true_props(solution$H, true_proportions)
plot_ptp_scatter(ptp)
plot_ptb_scatter(ptb)
```


#### Visualize the same as lines
```{r}
ptp_lines <- linseed::plotProportions(
  as.data.frame(ptp[[1]]),
  as.data.frame(ptp[[2]]),
  pnames = c("predicted", "true"),
  point_size = 1,
  line_size = 0.7
) + theme_bw(base_size = 12) + theme(
  legend.title = element_blank(),
  legend.position = "bottom",
  axis.title.x = element_blank(),
  axis.text.x = element_text(angle = 45, hjust = 1)
)

filename <- paste0(dir_to_save_fig, "ptp_lines.svg")
ggsave(
  filename,
  ptp_lines,
  width = 3.5,
  height = 3,
  device = svg
)
ptp_lines
```


### Select and visualize marker genes
```{r fig.height = 5, fig.width = 11}
markers  <- dso$get_marker_genes()
to_plot <- data_raw[unlist(markers), apply(true_proportions, 2, function(x) any(x == 1))]
dso$plot_projected(rownames(to_plot), use_dims = (2:3))
```


### More decorated markers plot (as it is in the paper)
#### Visualize with UMAP
```{r}
colors <- which_marker(rownames(fData(dso$st$data)), dso$st$marker_genes)
plot_projection_points(
  dso$st$proj,
  use_dims = NULL,
  spaces = c("X"),
  pt_size = 2,
  color = colors
) +
  scale_color_manual(values = colors_v[3:6],
                     na.value = adjustcolor("grey70", alpha.f = 0.7)) +
  labs(col = "Marker Cell Type", x = "UMAP1", y = "UMAP2") +
  theme_bw(base_family = "sans", base_size = 18) +
  theme(
    legend.position = "right",
    axis.ticks = element_blank(),
    axis.text = element_blank()
  )
```

#### Visualize original space
```{r}
colors <-
  which_marker(rownames(fData(dso$st$data)), dso$st$marker_genes)
plot_projection_points(
  dso$st$proj,
  use_dims = (3:4),
  spaces = c("X"),
  pt_size = 2,
  color = colors
) +
  scale_color_manual(values = colors_v[3:6],
                     na.value = adjustcolor("grey70", alpha.f = 0.7)) +
  labs(col = "Marker Cell Type", x = "UMAP1", y = "UMAP2") +
  theme_bw(base_family = "sans", base_size = 18) +
  theme(
    legend.position = "right",
    axis.ticks = element_blank(),
    axis.text = element_blank()
  )
```

### Nice 3D visualization of markers
#### Plot genes
```{r}
library(plotly)

n <- 2
dims <- dso$st$n_cell_types
X <- dso$st$solution_proj$X
colnames(X) <- paste0("R", c(1:dims))
Omega <- dso$st$solution_proj$Omega
colnames(Omega) <- paste0("S", c(1:dims))


l <- rep(list(1:nrow(X)), n)
permutations <- expand.grid(l)
indices <- c(t(as.matrix(permutations[permutations[,"Var1"] < permutations[, 'Var2'],] )))
X_points <- X[indices,]
Omega_points <- Omega[indices,]



colors <- which_marker(rownames(fData(dso$st$data)), dso$st$marker_genes)

font <- list(
  family = "Helvetica",
  size = 30)
ax_settings <- list(title = '',
                    autorange = T,
                    showgrid = T,
                    zeroline = F,
                    showline = T,
                    autotick = T,
                    ticks = '',
                    showticklabels = F,
                    titlefont = font)
to_plot <- as.data.frame(dso$st$proj$X)
to_plot$color <- as.character(colors)
to_plot$color[is.na(to_plot$color)] <- "not a marker"
fig_X <- plotly::plot_ly(
  to_plot,
  x = ~R2,
  y = ~R3,
  z = ~R4,
  color = ~color,
   colors = c(colors_v[3:6], "grey70"),
  opacity = 0.9,
  size = 1,
  type = "scatter3d",

  mode = "markers"
)  %>%
  plotly::layout(showlegend = F,
    scene = list(
      xaxis = c(title = "R2", ax_settings),
      yaxis = c(title = "R3", ax_settings),
      zaxis = c(title = "R4", ax_settings)
    )
  )  %>%  config(toImageButtonOptions = list(format = "svg", width = 600,
                                                              height = 600))


fig_X <- fig_X %>%  add_trace(x = X_points[,'R2'], y = X_points[,'R3'], z = X_points[,'R4'], color=NA,
            line = list(color = 'rgb(115, 115, 115)', width = 5))
fig_X
```

#### Plot samples

```{r}
font <- list(family = "Helvetica",
             size = 30)
ax_settings <- list(
  title = '',
  autorange = T,
  showgrid = T,
  zeroline = F,
  showline = T,
  autotick = T,
  ticks = '',
  showticklabels = F,
  titlefont = font
)
to_plot <- as.data.frame(dso$st$proj$Omega)
fig_Omega <- plotly::plot_ly(
  to_plot,
  x = ~ S2,
  y = ~ S3,
  z = ~ S4,
  color = "grey40",
  colors = c("navyblue"),
  opacity = 1,
  size = 1,
  type = "scatter3d",

  mode = "markers"
)  %>%
  plotly::layout(showlegend = FALSE,
                 scene = list(
                   xaxis = c(title = "S2", ax_settings),
                   yaxis = c(title = "S3", ax_settings),
                   zaxis = c(title = "S4", ax_settings)
                 ))  %>%  config(toImageButtonOptions = list(
                   format = "svg",
                   width = 600,
                   height = 600
                 ))


fig_Omega <-
  fig_Omega %>%  add_trace(
    x = Omega_points[, 'S2'],
    y = Omega_points[, 'S3'],
    z = Omega_points[, 'S4'],
    color = NA,
    opacity = 0.4,
    line = list(
      color = 'rgb(115, 115, 115)',
      opacity = 0.4,
      width = 5
    )
  )
fig_Omega
```


## Plot heatmap of marker genes expressions per component
```{r fig.height = 5, fig.width = 2}
markers  <- dso$get_marker_genes()

# Gess correct order of proportions column here
solution <- dso$finalize_solution()
reordered_H <- solution$H[unlist(guess_order(solution$H, true_proportions)), ]
sorted_predicted_cell_numbers <- strtoi(gsub("^.*_", "", rownames(reordered_H)))

to_plot <- data_raw[unlist(markers), apply(true_proportions, 2, function(x) any(x == 1))]

# Minmax scaling for visualization only
to_plot <- as.data.frame(apply(to_plot, 1, function(x) (x - min(x)) / (max(x) - min(x))))



#ht_opt$TITLE_PADDING = unit(c(30, 30), "points")
#ht_opt$COLUMN_ANNO_PADDING = unit(c(20, 20), "pt")
nrows <- dim(to_plot)[[1]]
ncols <- dim(to_plot)[[2]]
#  "Jurkat" "IM-9"   "Raji"   "THP-1"
factor_rows <- factor(rep(rownames(true_proportions), each=nrows/n_ct),
                      levels=c( "Jurkat", "IM-9" ,"Raji", "THP-1"))

factor_columns <- factor(rep(c(1:n_ct), each=ncols/n_ct), levels=sorted_predicted_cell_numbers)

marker_map <- Heatmap(
  to_plot,
  name = "W",
  cluster_rows = FALSE,
  cluster_columns = FALSE,
  show_column_names = FALSE,
  show_row_names = FALSE,
  show_heatmap_legend = F,
  column_split = factor_columns,
  row_split = factor_rows,
  use_raster = T,
  column_title_gp = gpar(
    fill = colors_v[3:6][sorted_predicted_cell_numbers],
    border = "black",
    fontsize = 20
  ),
  row_title_gp = gpar(fontsize = 20),
  column_title_side  = "bottom"
)

filename <- paste0(dir_to_save_fig, "heatmap.svg")
svg(filename, width=2, height=5)
draw(marker_map)
dev.off()
marker_map
```
